fix(logs): le projet ne compilait plus des qu'on passait logger.level a DEBUG
`on_wake_word_detected` avait `args: ['!lambda return wake_word;']`. Les entrees
de `args:` sont des expressions C++ BRUTES, pas des lambdas : le `!lambda` etait
recopie tel quel dans le code genere, d'ou

    ESP_LOGD("main", "DEBUG: WAKE WORD DETECTED (%s)", !lambda return wake_word;);
    error: 'lambda' was not declared in this scope

`wake_word` est un `std::string` (Trigger<std::string>), `%s` veut donc
`.c_str()`.

Le piege qui a permis a ce code mort de survivre : **tout `logger.log` est un
ESP_LOGD, donc du code jamais compile tant que `logger.level` reste a INFO.**
L'erreur n'apparaissait qu'en passant le projet en DEBUG — c'est-a-dire
exactement quand on en a besoin pour diagnostiquer quelque chose. Note laissee
sur `logger.level`.

Au passage, les 3 `-Wformat` de tab5_custom.cpp (`%x` avec un `uint32_t`, qui
est un `long unsigned int` sur cette cible) : cast explicite, la conversion est
l'identite et ces valeurs sont des couleurs RGB bornees a 0xFFFFFF.

Verifie : build complet a DEBUG **Successfully compiled program** (1 min 38,
`config_hash=0x472baff6`), 0 erreur et 0 `warning: format` ; flash 38,7 -> 39,0 %
(les corps des ESP_LOGD entrent enfin dans le binaire). Retour a INFO, build
`config_hash=0x0d3fa349`, OTA verifiee par
`device_info().compilation_time = 19:55:33` et sonnerie re-testee sur
l'appareil : 0 « Parent bus is busy ».

+++ Tab5/tab5_custom.cpp
@@ -477,8 +477,11 @@ void refresh_daily_forecast(WeatherDaySlot slots[], int page_index,
         // Tmin / Tmax colors
         uint32_t cmax = get_temperature_color(data.tmax);
         uint32_t cmin = get_temperature_color(data.tmin);
-        char buftx[64]; sprintf(buftx, "#%06x %.0f# / ", cmax, data.tmax);
-        char buftn[64]; sprintf(buftn, " #%06x %.0f# \xC2\xB0", cmin, data.tmin);
+        // `%x` attend un `unsigned int` ; `uint32_t` est un `long unsigned int`
+        // sur cette cible -> -Wformat. La conversion est l'identité (32 bits des
+        // deux côtés) et ces valeurs sont des couleurs RGB, bornées à 0xFFFFFF.
+        char buftx[64]; sprintf(buftx, "#%06x %.0f# / ", (unsigned) cmax, data.tmax);
+        char buftn[64]; sprintf(buftn, " #%06x %.0f# \xC2\xB0", (unsigned) cmin, data.tmin);
 
         lv_label_set_text(slot.max_lbl, data.est_passe ? "-- / " : buftx);
         lv_label_set_text(slot.min_lbl, data.est_passe ? "-- \xC2\xB0" : buftn);
@@ -544,7 +547,7 @@ void refresh_hourly_forecast(WeatherHourSlot slots[], int page_index,
         lv_label_set_text(slot.time_lbl, data.heure_texte.c_str());
         
         uint32_t c_t = get_temperature_color(data.temp);
-        char b_t[32]; sprintf(b_t, "#%06x %.0f#\xC2\xB0", c_t, data.temp);
+        char b_t[32]; sprintf(b_t, "#%06x %.0f#\xC2\xB0", (unsigned) c_t, data.temp);  // cf. -Wformat plus haut
         lv_label_set_text(slot.temp_lbl, b_t);
         lv_label_set_recolor(slot.temp_lbl, true);
         lv_obj_set_style_text_color(slot.temp_lbl, lv_color_hex(UIColor::TEXT_PRIMARY), LV_PART_MAIN);
